feat(frontend/student): reorganize hooks and update import paths for Forget and Reset Password components
